average_function/requirements.txt
flask==3.0.3
google-cloud-firestore==2.16.0
gunicorn==22.0.0
